fin calcul pouboire

+++ S03TipsCalculator/app/src/main/java/ca/qc/cstj/tipscalculator/domain/models/TipsCalculation.kt
@@ -4,7 +4,7 @@ import ca.qc.cstj.tipscalculator.core.Constants
 
 class TipsCalculation(private val subTotal: Double,private val tipsPercent: Int) {
 
-    val tipsAmount = subTotal * ( tipsPercent / Constants.PERCENT)
+    val tipsAmount = subTotal * ( tipsPercent.toFloat() / Constants.PERCENT)
     val TPS = subTotal * Constants.TPS
     val TVQ = subTotal * Constants.TVQ
     val total = subTotal + TPS + TVQ + tipsAmount
